Conflicts:
	CalDynam/bin/classes/com/example/caldynam/ExerciceActivity.class
	CalDynam/bin/classes/com/example/caldynam/MainActivity.class

+++ CalDynam/src/com/example/caldynam/AlimentationActivity.java
@@ -1,15 +1,12 @@
 package com.example.caldynam;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
-import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.DefaultHttpClient;
 import org.apache.http.params.BasicHttpParams;
 import org.json.JSONArray;
@@ -34,7 +31,7 @@ public class AlimentationActivity extends Activity implements OnClickListener {
 	private Button btnRechercheAliment, btnTerminerAliment;
 	private EditText edtRechercheAliment;
 	private TextView txtListeAliment;
-	private int totalIN;
+	private float totalIN;
 	
 	@Override
 	  protected void onCreate(Bundle savedInstanceState) {
